requirements.txt
faiss-cpu
numpy
tiktoken
sentence-transformers
torch
transformers
huggingface-hub
accelerate
pypdf
pdfminer.six